introduce format flag to group list
