+++ cmd/rrh/commands/testdata/broken.json
@@ -1,0 +1,1 @@
+{"last_modified":"2021-09-19T14:56:23+09:00","repositories":[],"groups":[],"relations":[]}
