+++ go.mod
@@ -1,6 +1,6 @@
 module github.com/tamada/rrh
 
-go 1.12
+go 1.16
 
 require (
 	github.com/armon/go-radix v1.0.0 // indirect
@@ -8,10 +8,12 @@ require (
 	github.com/emirpasic/gods v1.12.0 // indirect
 	github.com/go-playground/universal-translator v0.18.0 // indirect
 	github.com/gookit/color v1.1.7
+	github.com/karlseguin/jsonwriter v1.0.3 // indirect
 	github.com/leodido/go-urn v1.2.1 // indirect
 	github.com/mattn/go-isatty v0.0.8 // indirect
 	github.com/mitchellh/cli v1.0.0
 	github.com/mitchellh/go-homedir v1.1.0
+	github.com/olekukonko/tablewriter v0.0.5 // indirect
 	github.com/posener/complete v1.2.1 // indirect
 	github.com/skratchdot/open-golang v0.0.0-20200116055534-eef842397966
 	github.com/spf13/cobra v1.2.1
